requirements.txt
absl-py==0.11.0
aiohttp==3.7.4
appdirs==1.4.3
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
bravado==11.0.3
bravado-core==5.17.0
CacheControl==0.12.6
cachetools==4.2.1
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
entrypoints==0.3
fastapi==0.63.0
filelock==3.0.12
flatbuffers==1.12
fsspec==0.8.7
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.14
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
gym==0.18.0
h11==0.12.0
h5py==2.10.0
html5lib==1.0.1
hyperopt==0.1.2
idna==2.8
ipaddr==2.2.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
json-tricks==3.15.5
jsonpointer==2.0
jsonref==0.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lockfile==0.12.2
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
monotonic==1.5
msgpack==0.6.2
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
netifaces==0.10.9
networkx==2.5
nni==2.1
notebook==6.2.0
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.3
pandas==1.2.2
pandocfilters==1.4.3
parso==0.8.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.14.3
prettytable==2.1.0
progress==1.5
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.15.5
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.1
pyglet==1.5.0
Pygments==2.8.0
PyJWT==2.0.1
pymongo==3.11.3
pyparsing==2.4.6
pyrsistent==0.17.3
python-dateutil==2.8.1
PythonWebHDFS==0.2.3
pytoml==0.1.21
pytorch-lightning==1.2.1
pytz==2021.1
PyYAML==5.3.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
requests==2.22.0
requests-oauthlib==1.3.0
responses==0.13.2
retrying==1.3.3
rfc3987==1.3.8
rsa==4.7.2
ruamel.yaml==0.17.2
ruamel.yaml.clib==0.2.2
schema==0.7.4
scikit-learn==0.24.1
scipy==1.6.1
Send2Trash==1.5.0
simplejson==3.17.2
six==1.15.0
smmap==3.0.5
starlette==0.13.6
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
torch==1.7.1
tornado==6.1
tqdm==4.58.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websocket-client==0.58.0
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.6.3
